en: 完成添加sbt弹出层样式
