+++ src/assets/styles/mobile/view-style/cert.scss
@@ -162,4 +162,144 @@ $container-width: 90vw;
             }
         }
     }
+    .Cert-addsbt{
+        position: relative;
+        .back{
+            margin-bottom: 1.4375rem;
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+            .icon{
+                width: 1.875rem;
+                height: 1.875rem;
+                display: flex;
+                align-items: center;
+                justify-content: center;
+                background-color: #EDF1F4;
+                border-radius: .125rem;
+            }
+        }
+        .title{
+            font-size: 1.125rem;
+            font-weight: 700;
+            line-height: 1.625rem;
+        }
+        .search{
+            margin-top: .9375rem;
+            .confirm{
+                min-width: 5rem;
+                height: 1.875rem;
+                padding: 0;
+                border-radius: .9375rem;
+                font-size: .875rem;
+                color: #9E9E9E;
+                line-height: 1.625rem;
+                border: .125rem solid #BABABA;
+                position: absolute;
+                right: 0;
+                top: 0;
+            }
+            .inner{
+                width: 100%;
+                height: 2.1875rem;
+                border: .125rem solid #BABABA;
+                border-radius: 1.125rem;
+                display: flex;
+                align-items: center;
+                .ant-select-selection-item{
+                    text-overflow: clip;
+                }
+                .input{
+                    width: 14.375rem;
+                    height: 100%;
+                    border-left: .125rem solid #BABABA;
+                    position: relative;
+                    input{
+                        padding-left: 2.5rem;
+                    }
+                    .icon{
+                        position: absolute;
+                        left: .875rem;
+                        top: 0rem;
+                        color: #BABABA;
+                        font-size: 1.125rem;
+                        line-height: 1.9375rem;
+                    }
+                }
+            }
+            .ant-select-selector{
+                padding: 0 .5rem;
+                border: none;
+            }
+            .label{
+                font-size: .75rem;
+                color: #9E9E9E;
+                width: 3.125rem;
+                overflow: hidden;
+                text-overflow: ellipsis;
+            }
+        }
+        .ant-divider{
+            display: none;
+        }
+        .content{
+            padding: 1.25rem 0;
+            height: 28.125rem;
+            overflow-y: scroll;
+            .list-content{
+                flex-wrap: wrap;
+                display: flex;
+                gap: .75rem .375rem;
+            }
+            .box-active, .box:hover{
+                .checkbox{
+                    display: block !important;
+                }
+            }
+            .box{
+                width: 10.4375rem;
+                height: 13.75rem;
+                margin-left: .25rem;
+                position: relative;
+                z-index: 1;
+
+                .checkbox{
+                    width: 2rem;
+                    height: 2rem;
+                    border: .0625rem solid #CECECE;
+                    position: absolute;
+                    right: 0;
+                    bottom: 0;
+                    display: none;
+                    text-align: center;
+                    line-height: 2rem;
+                    color: #07A35A;
+                    border-radius: .1875rem;
+                }
+                display: flex;
+                flex-direction: column;
+                align-items: center;
+                text-align: center;
+                box-shadow: 0 0 .3125rem -0.0625rem #999;
+                border-radius: .375rem;
+                .img{
+                    width: 6.6875rem;
+                    height: 6.6875rem;
+                    margin-top: 1.25rem;
+                }
+                p{
+                    width: 100%;
+                    padding-left: .625rem;
+                    padding-right: .625rem;
+                    color: #17151C;
+                    font-size: .5625rem;
+                    font-weight: 500;
+                    line-height: .8125rem;
+                    white-space: nowrap;  
+                    text-overflow:ellipsis; 
+                    overflow:hidden;
+                }
+            }
+        }
+    }
 }
