+++ src/components/Cert/Nfts.js
@@ -9,11 +9,15 @@ import ModalAddSbt from "./ModalAddSbt";
 
 export default function CertNfts(props) {
     
-    const { account, changeContractId, total, isMe, nftlist: list, isMobile } = props;
+    const { account, changeContractId, total, isMe, nftlist: list, isMobile, goAddSbt } = props;
     const [isModalOpen, setIsModalOpen] = useState(false);
     let [selectItem, setSelectItem] = useState(isMobile ? null : 0);
 
     const show = () => {
+        if (isMobile) {
+            goAddSbt()
+            return
+        }
         setIsModalOpen(true);
     }
